Final Commit for Stickies 1.0
Stickies 1.0 is complete. If there are any glitches found I will update accordingly. I added realtime updates without relying on client-refreshing. I also added a shop mechanism which does not currently work because PayPal is very confusing. Next on the agenda is to make a logo for Stickies. After I make that, I will make a commit to all of my index.html files to make it the tab icon.